Added flavoring
Overloading/Adding to a source ressource file before export is now
implemented. Therefor a new parameter *--flavor filename.json* was added
to select the file with the contents for the flavoring process. During
flavoring strings with same keys as in the flavor file will be replaced,
including all given attributes from flavor, and new strings will simply
be added.